Fix html, add styles for the 1 and 2 part of page
index.html structure was incorrect, because of this the correct styles could not be applied.
I had to change the structure. I also added the correct styles for the first and second part of the page,
so that the site would display correctly, as specified in the terms of reference.